refactor(TableDisplay.js): Moved table display into a table directory that will be used for shared tables in the future
