+++ ui/src/main/js/dynamic/loaded/tasks/TaskManagement.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import ConfigurationLabel from 'component/common/ConfigurationLabel';
-import TableDisplay from 'field/TableDisplay';
+import TableDisplay from 'field/table/TableDisplay';
 import ReadOnlyField from 'field/ReadOnlyField';
 import { fetchTasks } from 'store/actions/tasks';
 import { TASK_MANAGEMENT_INFO } from 'global/components/task/TaskManagementModel';
